Replace CryptoJS with a simlpe XOR.
Using a Crypto-JS library was offering a false sense of security when
using a hardcoded password, while increasing the file-size over 4 times.
Let us use a simple XOR cipher that would just obfuscate the data until
we initialize the PosMeta object.

It is simpler, faster, and achieves the same effect.